+++ src/game/tests.rs
@@ -329,6 +329,7 @@ mod vintage_cube_graveyard;
 mod vintage_cube_gut;
 mod vintage_cube_hand_attack;
 mod vintage_cube_haywire_mite;
+mod vintage_cube_headliner_scarlett;
 mod vintage_cube_horizon_land;
 mod vintage_cube_infect;
 mod vintage_cube_ivora;